create components folder and extract todo page
